requirements.txt
nonebot2==2.0.0b3
nonebot-adapter-onebot==2.1.1
fastapi==0.78.0
pydantic==1.9.1
APScheduler~=3.9.1
pytz~=2022.1
aiohttp~=3.8.1
SQLAlchemy~=1.4.37
asyncmy~=0.2.5
aiomysql~=0.1.1
aiofiles==0.8.0
ujson~=5.3.0
beautifulsoup4~=4.11.1
bs4~=0.0.1
lxml~=4.9.0
numpy~=1.22.4
matplotlib~=3.5.2
Pillow~=9.1.1
imageio~=2.19.3
msgpack~=1.0.4
pycryptodome~=3.14.1
py7zr~=0.18.9
zhconv~=1.4.3
rapidfuzz~=2.0.11
emoji~=1.7.0
mcstatus